add normalization

+++ utils.py
@@ -35,7 +35,7 @@ def train_epoch(model, dataloader, dataset, optimizer, loss, device):
     for batch in tqdm(dataloader, desc="Training"):
         optimizer.zero_grad()
         images, texts, labels = batch
-        images = images.to(device) # float32 to(torch.float32)
+        images = images.to(torch.float32).to(device) # float32 to(torch.float32)
         texts = texts.to(device)
         labels = labels.to(device)
 
@@ -66,7 +66,7 @@ def validate(model, dataloader, dataset, loss, device):
     with torch.no_grad():
         for batch in tqdm(dataloader, desc="Validation"):
             images, texts, labels = batch
-            images = images.to(device)
+            images = images.to(torch.float32).to(device)
             texts = texts.to(device)
             labels = labels.to(device)
